Allow for request headers that are not dictionaries.
The requests library implements a `CaseInsensitiveDict` structure that is
used for headers instead of a dictionary in some request models. This
causes problems in `jaeger-client-python` codecs which test if the
carrier param is an instance of `dict`.

This change checks for headers of this type and converts them to `type
dict`.